+++ lib/ruby-next/language/rewriters/in_pattern.rb
@@ -16,39 +16,7 @@ module RubyNext
 
         # Make case-match no-op
         def on_case_match(node)
-          node
-        end
-
-        def on_match_pattern_p(node)
-          context.track! self
-
-          @deconstructed_keys = {}
-          @predicates = Predicates::Noop.new
-
-          matchee =
-            s(:begin, s(:lvasgn, MATCHEE, node.children[0]))
-
-          pattern =
-            locals.with(
-              matchee: MATCHEE,
-              arr: MATCHEE_ARR,
-              hash: MATCHEE_HASH
-            ) do
-              send(
-                :"#{node.children[1].type}_clause",
-                node.children[1]
-              )
-            end
-
-          node.updated(
-            :and,
-            [
-              matchee,
-              pattern
-            ]
-          ).tap do |new_node|
-            replace(node.loc.expression, inline_blocks(unparse(new_node)))
-          end
+          process_regular_node(node)
         end
       end
     end